move mainnet warning modal higher up the dom tree
